--- packages/pre-bundle-new-url/src/vite.ts
@@ -1,21 +1,0 @@
-import type { Plugin } from "vite";
-import { esbuildPluginNewUrl, esbuildPluginWorkerNewUrl } from "./esbuild";
-
-export function vitePluginPreBundleNewUrl(): Plugin {
-  return {
-    name: "pre-bundle-new-url",
-    config(config, _env) {
-      config.cacheDir;
-      return {
-        optimizeDeps: {
-          esbuildOptions: {
-            plugins: [esbuildPluginWorkerNewUrl(), esbuildPluginNewUrl()],
-          },
-        },
-      };
-    },
-    configResolved(config) {
-      config.cacheDir;
-    },
-  };
-}
